// src/entities/payment/model/Payment.ts
export interface PaymentItem {
  id?: string;
  s_idx: number;
  ss_idx: number;
  b_idx: number;
  brand_code: string;
  amount: number;
  b_type: '지정지점' | '전지점';
  title: string;
  brand: string;
  branchType: string;
  duration: number;
  price: number;
  originalPrice?: number;
  sessions: number;
  name?: string;
  type: 'membership' | 'additional';
  // 추가관리 전용 필드
  mp_idx?: number;
  r_date?: string;
  r_memo?: string;
  branchName?: string;
}